Simplify login method in user controller to return just a 200
- Any messaging can be defined on the client side (unless it's a server
related issue). This decouples our client and server further (the UI
previously had to understand the structure of the json response
returned). :) Context can be inferred through the HTTP status code
returned.
- Fix tests using stubs for login
- Move test to __tests__ folder
