+++ server/db/mongo/controllers/users.js
@@ -9,15 +9,13 @@ export function login(req, res, next) {
   passport.authenticate('local', (authErr, user, info) => {
     if (authErr) return next(authErr);
     if (!user) {
-      return res.status(401).json({ message: info.message });
+      return res.sendStatus(401);
     }
     // Passport exposes a login() function on req (also aliased as
     // logIn()) that can be used to establish a login session
     return req.logIn(user, (loginErr) => {
-      if (loginErr) return res.status(401).json({ message: loginErr });
-      return res.status(200).json({
-        message: 'You have been successfully logged in.'
-      });
+      if (loginErr) return res.sendStatus(401);
+      return res.sendStatus(200);
     });
   })(req, res, next);
 }
